requirements.txt
AfricasTalkingGateway==1.6
